+++ src/Makefile.am
@@ -303,7 +303,6 @@ libBGL_server_a_SOURCES = \
   index/blockfilterindex.cpp \
   index/txindex.cpp \
   init.cpp \
-  interfaces/chain.cpp \
   miner.cpp \
   net.cpp \
   net_processing.cpp \
